Move WizardBar out of WizardPage
