+++ packages/ubuntu_desktop_installer/lib/pages/turn_off_bitlocker/turn_off_bitlocker_page.dart
@@ -95,10 +95,9 @@ class TurnOffBitLockerPage extends StatelessWidget {
           ],
         ),
       ),
-      singleLeading: true,
-      actions: [
-        WizardAction.back(context),
-      ],
+      bottomBar: WizardBar(
+        leading: WizardAction.back(context),
+      ),
     );
   }
 }
