+++ packages/ubuntu_desktop_installer/lib/pages/turn_off_rst/turn_off_rst_page.dart
@@ -96,10 +96,9 @@ class TurnOffRSTPage extends StatelessWidget {
             ],
           ),
         ),
-        singleLeading: true,
-        actions: <WizardAction>[
-          WizardAction.back(context),
-        ],
+        bottomBar: WizardBar(
+          leading: WizardAction.back(context),
+        ),
       ),
     );
   }
