+++ packages/ubuntu_wizard/lib/widgets.dart
@@ -5,4 +5,5 @@ export 'src/widgets/option_card.dart';
 export 'src/widgets/override_mouse_cursor.dart';
 export 'src/widgets/password_strength_label.dart';
 export 'src/widgets/validated_form_field.dart';
+export 'src/widgets/wizard_bar.dart';
 export 'src/widgets/wizard_page.dart';
